add social blurb props
