+++ src/app/socials/page.tsx
@@ -1,27 +1,12 @@
-import SocialBanner, { SocialBlurb } from "@/components/socials/Social";
-import instagram from "@/public/social/instagram.webp";
-import email from "@/public/social/email.webp";
+import { SocialBanner, SocialPage } from "@/components/socials/Social";
 
-const SocialPage = () => {
+const SocialPageWrapper = () => {
   return (
     <div>
       <SocialBanner />
-      <div>
-        <SocialBlurb
-          title="Follow us on Instagram!"
-          description="Keep up with the details about upcoming new events!"
-          image={instagram}
-          link=""
-        />
-        <SocialBlurb
-          title="Reach Out!"
-          description="Here's our email if you need to contact us!"
-          image={email}
-          link=""
-        />
-      </div>
+      <SocialPage />
     </div>
   );
 };
 
-export default SocialPage;
+export default SocialPageWrapper;
