+++ src/components/socials/Social.tsx
@@ -1,36 +1,32 @@
-import Image, { StaticImageData } from "next/image";
+import SocialBlurb from "./SocialsBlurbProps";
 import socialsimage from "@/public/social/socialsimage.webp";
+import instagram from "@/public/social/instagram.webp";
+import email from "@/public/social/email.webp";
+import Image from "next/image";
 
-const SocialBanner = () => {
+export const SocialBanner = () => {
   return (
     <div className="w-full">
       <Image src={socialsimage} alt="Socials Banner" />
     </div>
   );
 };
-interface BlurbProps {
-  title: string;
-  description: string;
-  image: StaticImageData;
-  link: string;
-}
 
-const SocialBlurb: React.FC<BlurbProps> = ({
-  title,
-  description,
-  image,
-  link,
-}) => {
+export const SocialPage = () => {
   return (
-    <link href={link}>
-      <div>
-        <Image src={image} alt={title} />
-        <h2>{title}</h2>
-        <p>{description}</p>
-      </div>
-    </link>
+    <div className="flex justify-center">
+      <SocialBlurb
+        title="Follow us on Instagram!"
+        description="Keep up with the details about upcoming new events!"
+        image={instagram}
+        link=""
+      />
+      <SocialBlurb
+        title="Reach Out!"
+        description="Here's our email if you need to contact us!"
+        image={email}
+        link=""
+      />
+    </div>
   );
 };
-
-export default SocialBanner;
-export { SocialBlurb };
